+++ java-client/src/main/java/io/sphere/internal/CustomerServiceImpl.java
@@ -27,32 +27,28 @@ public class CustomerServiceImpl extends ProjectScopedAPI implements CustomerSer
         this.requestFactory = requestFactory;
     }
 
-    /** {@inheritDoc}  */
-    public FetchRequest<Customer> byId(String id) {
+    @Override public FetchRequest<Customer> byId(String id) {
         return requestFactory.createFetchRequest(
                 endpoints.customers.byId(id),
                 Optional.<ApiMode>absent(),
                 new TypeReference<Customer>() {});
     }
 
-    /** {@inheritDoc}  */
-    public FetchRequest<Customer> byToken(String token) {
+    @Override public FetchRequest<Customer> byToken(String token) {
         return requestFactory.createFetchRequest(
                 endpoints.customers.byToken(token),
                 Optional.<ApiMode>absent(),
                 new TypeReference<Customer>() {});
     }
 
-    /** {@inheritDoc}  */
-    public QueryRequest<Customer> all() {
+    @Override public QueryRequest<Customer> all() {
         return requestFactory.createQueryRequest(
                 endpoints.customers.root(),
                 Optional.<ApiMode>absent(),
                 new TypeReference<QueryResult<Customer>>() {});
     }
 
-    /** {@inheritDoc}  */
-    public FetchRequest<AuthenticatedCustomerResult> byCredentials(String email, String password) {
+    @Override public FetchRequest<AuthenticatedCustomerResult> byCredentials(String email, String password) {
         return requestFactory.createFetchRequestWithErrorHandling(
                 endpoints.customers.login(email, password),
                 Optional.<ApiMode>absent(),
@@ -60,25 +56,21 @@ public class CustomerServiceImpl extends ProjectScopedAPI implements CustomerSer
                 new TypeReference<AuthenticatedCustomerResult>() {});
     }
 
-    /** {@inheritDoc}  */
-    public CommandRequest<Customer> signup(String email, String password, String firstName, String lastName, String middleName, String title) {
+    @Override public CommandRequest<Customer> signup(String email, String password, String firstName, String lastName, String middleName, String title) {
         return createCommandRequest(
                 endpoints.customers.root(),
                 new CustomerCommands.CreateCustomer(email, password, firstName, lastName, middleName, title));
     }
 
-    /** {@inheritDoc}  */
-    public CommandRequest<AuthenticatedCustomerResult> signupWithCart(
-            String email, String password, String firstName, String lastName, String middleName, String title, String cartId, int cartVersion)
-    {
+    @Override public CommandRequest<AuthenticatedCustomerResult> signupWithCart(
+            String email, String password, String firstName, String lastName, String middleName, String title, String cartId, int cartVersion) {
         return requestFactory.createCommandRequest(
             endpoints.customers.signupWithCart(),
             new CustomerCommands.CreateCustomerWithCart(email, password, firstName, lastName, middleName, title, cartId, cartVersion),
             new TypeReference<AuthenticatedCustomerResult>() {});
     }
 
-    /** {@inheritDoc}  */
-    public CommandRequest<Optional<Customer>> changePassword(String customerId, int customerVersion, String currentPassword, String newPassword) {
+    @Override public CommandRequest<Optional<Customer>> changePassword(String customerId, int customerVersion, String currentPassword, String newPassword) {
         return requestFactory.createCommandRequestWithErrorHandling(
                 endpoints.customers.changePassword(),
                 new CustomerCommands.ChangePassword(customerId, customerVersion, currentPassword, newPassword),
@@ -86,47 +78,42 @@ public class CustomerServiceImpl extends ProjectScopedAPI implements CustomerSer
                 new TypeReference<Customer>() {});
     }
 
-    /** {@inheritDoc}  */
-    public CommandRequest<Customer> update(String customerId, int customerVersion, CustomerUpdate customerUpdate) {
+    @Override public CommandRequest<Customer> update(String customerId, int customerVersion, CustomerUpdate customerUpdate) {
         return createCommandRequest(
                 endpoints.customers.byId(customerId),
                 new UpdateCommand<CustomerCommands.CustomerUpdateAction>(customerVersion, customerUpdate));
     }
 
-    /** {@inheritDoc}  */
-    public CommandRequest<CustomerToken> createPasswordResetToken(String email) {
+    @Override public CommandRequest<CustomerToken> createPasswordResetToken(String email) {
         return createCustomerTokenCommandRequest(
                 endpoints.customers.createPasswordResetToken(),
                 new CustomerCommands.CreatePasswordResetToken(email));
     }
 
-    /** {@inheritDoc}  */
-    public CommandRequest<Customer> resetPassword(String customerId, int customerVersion, String tokenValue, String newPassword) {
+    @Override public CommandRequest<Customer> resetPassword(String customerId, int customerVersion, String token, String newPassword) {
         return createCommandRequest(
                 endpoints.customers.resetPassword(),
-                new CustomerCommands.ResetCustomerPassword(customerId, customerVersion, tokenValue, newPassword));
+                new CustomerCommands.ResetCustomerPassword(customerId, customerVersion, token, newPassword));
     }
 
-    /** {@inheritDoc}  */
-    public CommandRequest<CustomerToken> createEmailVerificationToken(String customerId, int customerVersion, int ttlMinutes) {
+    @Override public CommandRequest<CustomerToken> createEmailVerificationToken(String customerId, int customerVersion, int ttlMinutes) {
         return createCustomerTokenCommandRequest(
                 endpoints.customers.createEmailVerificationToken(),
                 new CustomerCommands.CreateEmailVerificationToken(customerId, customerVersion, ttlMinutes));
     }
 
-    /** {@inheritDoc}  */
-    public CommandRequest<Customer> confirmEmail(String customerId, int customerVersion, String tokenValue) {
+    @Override public CommandRequest<Customer> confirmEmail(String customerId, int customerVersion, String token) {
         return createCommandRequest(
                 endpoints.customers.confirmEmail(),
-                new CustomerCommands.VerifyCustomerEmail(customerId, customerVersion, tokenValue));
+                new CustomerCommands.VerifyCustomerEmail(customerId, customerVersion, token));
     }
 
-    /** Helper to save some repetitive code in this class. */
+    /** Helper to save some repetitive code. */
     private CommandRequest<Customer> createCommandRequest(String url, Command command) {
         return requestFactory.createCommandRequest(url, command, new TypeReference<Customer>() {});
     }
 
-    /** Helper to save some repetitive code in this class. */
+    /** Helper to save some repetitive code. */
     private CommandRequest<CustomerToken> createCustomerTokenCommandRequest(String url, Command command) {
         return requestFactory.createCommandRequest(url, command, new TypeReference<CustomerToken>() {});
     }